* Types 'uint64_t' removed to avoid '(long long)' casts in printf's.

+++ src/libcpuarch/uop.c
@@ -103,7 +103,7 @@ void uop_lnlist_check_if_ready(struct linked_list_t *uop_list)
 			continue;
 		uop->ready = 1;
 		esim_debug("uop action=\"update\", core=%d, seq=%lld, ready=1\n",
-			uop->core, (long long) uop->di_seq);
+			uop->core, uop->di_seq);
 	}
 }
